# Reduced memory allocations by user.ExtractFromGRPCRequest()